Testing a theory about having two suites in the same test file.
